fix datatype - int should be int...

+++ src/main/java/org/drizzle/jdbc/internal/mysql/MySQLType.java
@@ -50,7 +50,7 @@ public class MySQLType implements DataType {
         DECIMAL(java.sql.Types.DECIMAL, Double.class),
         TINY(java.sql.Types.SMALLINT, Short.class),
         SHORT(java.sql.Types.SMALLINT, Short.class),
-        LONG(java.sql.Types.BIGINT, Long.class),
+        LONG(java.sql.Types.INTEGER, Long.class),
         FLOAT(java.sql.Types.FLOAT, Float.class),
         DOUBLE(java.sql.Types.DOUBLE, Double.class),
         NULL(java.sql.Types.NULL, null),
